+++ app/views/layout.erb
@@ -16,6 +16,16 @@
 
     <% if current_user %>
     Hello, <%= current_user.username %>
+    <form method='post' action='/sessions'>
+      <input type='hidden' name='_method' value='delete'>
+      <input type='submit' value='Sign out'>
+    </form>
+    <% end %>
+
+    <% if flash[:notice] %>
+    <div id='notice'>
+      <%= flash[:notice] %>
+    </div>
     <% end %>
 
     <%= yield %>
