+++ spec/spec_helper.rb
@@ -8,8 +8,8 @@ require 'capybara'
 require 'capybara/rspec'
 require 'rspec'
 require 'database_cleaner'
-require 'web_helpers'
 require './app/data_mapper_setup.rb'
+require_relative 'helpers/session'
 
 Capybara.app = Chitter
 # hmm - should be generated by `rspec --init` directly rather than copy paste ...
@@ -44,6 +44,8 @@ RSpec.configure do |config|
   config.after(:each) do
     DatabaseCleaner.clean
   end
+
+  config.include SessionHelpers
 
   config.include Capybara::DSL
   # rspec-expectations config goes here. You can use an alternate
